requirements.txt
alpaca-py==0.6.0
certifi==2022.9.24
charset-normalizer==2.1.1
distlib==0.3.6
filelock==3.8.0
idna==3.4
msgpack==1.0.4
numpy==1.23.4
pandas==1.5.0
pipenv==2022.10.12
platformdirs==2.5.2
pydantic==1.10.2
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.4
requests==2.28.1
six==1.16.0
sseclient-py==1.7.2
typing_extensions==4.4.0
urllib3==1.26.12
websockets==10.3
